requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
_pytorch_select=0.1=cpu_0
alsa-lib=1.2.3.2=h166bdaf_0
blas=1.0=mkl
bzip2=1.0.8=h7f98852_4
c-ares=1.18.1=h7f98852_0
ca-certificates=2022.4.26=h06a4308_0
cairo=1.16.0=h6cf1ce9_1008
certifi=2022.6.15=py38h06a4308_0
cudatoolkit=11.0.221=h6bb024c_0
dbus=1.13.6=h48d8840_2
expat=2.4.8=h27087fc_0
ffmpeg=4.3.2=hca11adc_0
fontconfig=2.14.0=h8e229c2_0
freetype=2.11.0=h70c0345_0
gettext=0.19.8.1=h0b5b191_1005
giflib=5.2.1=h7b6447c_0
glib=2.68.4=h9c3ff4c_0
glib-tools=2.68.4=h9c3ff4c_0
gmp=6.2.1=h58526e2_0
gnutls=3.6.13=h85f3911_1
graphite2=1.3.13=h58526e2_1001
gst-plugins-base=1.18.5=hf529b03_0
gstreamer=1.18.5=h76c114f_0
harfbuzz=2.9.1=h83ec7ef_1
hdf5=1.10.6=hb1b8bf9_0
icu=68.2=h9c3ff4c_0
intel-openmp=2021.4.0=h06a4308_3561
jasper=1.900.1=h07fcdf6_1006
jpeg=9e=h166bdaf_1
keyutils=1.6.1=h166bdaf_0
krb5=1.19.3=h3790be6_0
lame=3.100=h7f98852_1001
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
libblas=3.9.0=12_linux64_mkl
libcblas=3.9.0=12_linux64_mkl
libclang=11.1.0=default_ha53f305_1
libcurl=7.79.1=h2574ce0_1
libedit=3.1.20191231=he28a2e2_2
libev=4.33=h516909a_1
libevent=2.1.10=h9b69904_4
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libglib=2.68.4=h3e27bee_0
libgomp=11.2.0=h1234567_1
libiconv=1.16=h516909a_0
liblapack=3.9.0=12_linux64_mkl
liblapacke=3.9.0=12_linux64_mkl
libllvm11=11.1.0=hf817b99_2
libnghttp2=1.43.0=h812cca2_1
libogg=1.3.4=h7f98852_1
libopencv=4.5.1=py38h703c3c0_0
libopus=1.3.1=h7f98852_1
libpng=1.6.37=hbc83047_0
libpq=13.3=hd57d9b9_0
libssh2=1.10.0=ha56f1ee_2
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.1.0=h2733197_1
libuuid=2.32.1=h7f98852_1000
libuv=1.40.0=h7b6447c_0
libvorbis=1.3.7=h9c3ff4c_0
libwebp=1.2.0=h89dd481_0
libwebp-base=1.2.0=h7f98852_2
libxcb=1.13=h7f98852_1004
libxkbcommon=1.0.3=he3ba5ed_0
libxml2=2.9.12=h72842e0_0
lz4-c=1.9.3=h295c915_1
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
mysql-common=8.0.25=ha770c72_0
mysql-libs=8.0.25=h935591d_0
ncurses=6.3=h7f8727e_2
nettle=3.6=he412f7d_0
ninja=1.10.2=h06a4308_5
ninja-base=1.10.2=hd09550d_5
nspr=4.32=h9c3ff4c_1
nss=3.69=hb5efdd6_1
numpy=1.22.3=py38he7a7128_0
numpy-base=1.22.3=py38hf524024_0
opencv=4.5.1=py38h578d9bd_0
openh264=2.1.1=h780b84a_0
openssl=1.1.1o=h7f8727e_0
pcre=8.45=h9c3ff4c_0
pillow=9.0.1=py38h22f2fdc_0
pip=21.2.4=py38h06a4308_0
pixman=0.40.0=h36c2ea0_0
pthread-stubs=0.4=h36c2ea0_1001
py-opencv=4.5.1=py38h81c977d_0
python=3.8.13=h12debd9_0
python_abi=3.8=2_cp38
pytorch=1.7.1=py3.8_cuda11.0.221_cudnn8.0.5_0
pyyaml=6.0=py38h7f8727e_1
qt=5.12.9=hda022c4_4
readline=8.1.2=h7f8727e_1
scipy=1.7.3=py38hc147768_0
setuptools=61.2.0=py38h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.38.5=hc218d9a_0
tk=8.6.12=h1ccaba5_0
torchaudio=0.7.2=py38
torchvision=0.8.2=cpu_py38ha229d99_0
tqdm=4.64.0=py38h06a4308_0
typing_extensions=4.1.1=pyh06a4308_0
wheel=0.37.1=pyhd3eb1b0_0
x264=1!161.3030=h7f98852_1
xorg-kbproto=1.0.7=h7f98852_1002
xorg-libice=1.0.10=h7f98852_0
xorg-libsm=1.2.3=hd9c2040_1000
xorg-libx11=1.7.2=h7f98852_0
xorg-libxau=1.0.9=h7f98852_0
xorg-libxdmcp=1.1.3=h7f98852_0
xorg-libxext=1.3.4=h7f98852_1
xorg-libxrender=0.9.10=h7f98852_1003
xorg-renderproto=0.11.1=h7f98852_1002
xorg-xextproto=7.3.0=h7f98852_1002
xorg-xproto=7.0.31=h7f98852_1007
xz=5.2.5=h7f8727e_1
yaml=0.2.5=h7b6447c_0
zlib=1.2.12=h7f8727e_2
zstd=1.4.9=haebb681_0
